Extracting VM and UI common code
